[0.2.0] Breaking change: no more `isLoadingMore` -- use a promise instead
`onLoadMore` has been replaced by `onLoadMoreAsync`. Pass in an async function, and InfiniteScrollView will keep track of whether or not is it's loading more content.

There is also support for an error indicator if we failed to load more content. It is passed a function you can call to retry the load (it will call `onLoadMoreAsync` again). You may let the user tap your error indicator to retry loading.