Enable canLoadMore to be a boolean or a callback
For an app I am building it's easier to evaluate canLoadMore dynamically so that I can factor in time when figuring out whether or not to load the next page.

I've expanded the code here so that if props.canLoadMore is a function, it will call the function and use its result.  It's backward compatible with the previous code, so props.canLoadMore can still be a boolean.